Setup repo to use docker-compose
The repo is now very very easy to install on any computer of any kind (currently writing this from Gentoo, works fine on Ubuntu, and probably works on Windows and Mac). 

Also made a minor change to the login code where it tells if you're already logged in when trying to log in again.