Fixed incorrect login error + faster login
Just made it so that when the script loops through the entire database and doesn't find the user, it says that it can't find the user specified (instead of the previous accidental behavior of just not saying anything). Also made the script exit when it finds a valid user and adds session variables, which increases efficiency.

Also removed Python virtual environment, it's extremely unnecessary and has no reason for being in the repo anymore.